Many changes
CRUD Operations,
RecyclerView Format updated,
Input Fields Updated and Styled,
Additional Helper Class created

+++ app/src/main/java/com/example/budgetingapp/models/MonthlyBudget.java
@@ -5,6 +5,9 @@ import java.math.RoundingMode;
 import java.time.YearMonth;
 import java.util.ArrayList;
 
+/**
+ * Currently unused
+ */
 public class MonthlyBudget {
     private ArrayList<BudgetLine> budgetLines;
     private YearMonth month;
@@ -14,7 +17,7 @@ public class MonthlyBudget {
     public BigDecimal monthlyTotal() {
         BigDecimal total = new BigDecimal(0);
         for(BudgetLine l : budgetLines) {
-            total = total.add(BigDecimal.valueOf(Double.valueOf(l.getAmount())));
+            total = total.add(l.getAmount());
         }
         total = total.setScale(2, RoundingMode.HALF_UP);
         return total;
